@@ -34,25 +34,25 @@ providers.forEach((provider) => {
 
   provider.on('connection', (client) => {
     if (isIntern) {
-      client.subscribe('n.add', (msg, evt) => {
+      client.subscribe('n.add', (body, frame) => {
         if (isSeed) {
-          provider.broadcast('n.add', msg);
-        } else provider.connect(evt.remote);
+          provider.broadcast('n.add', body);
+        } else provider.connect(frame.remote);
       });
-      client.subscribe('n.evt', (msg, evt) => {
+      client.subscribe('n.evt', (body, frame) => {
         providers.forEach((_provider) => {
           if (_provider.label === 'external') {
-            _provider.broadcast('r.evt', msg);
+            _provider.broadcast('r.evt', body);
           }
         });
       });
     } else {
-      client.subscribe('c.evt', (msg, evt) => {
+      client.subscribe('c.evt', (body, frame) => {
         providers.forEach((_provider) => {
           if (_provider.label === 'internal') {
-            _provider.broadcast('n.evt', msg);
+            _provider.broadcast('n.evt', body);
           } else {
-            _provider.broadcast('r.evt', msg);
+            _provider.broadcast('r.evt', body);
           }
         });
       });
