@@ -1,14 +1,12 @@
 /* Requires ------------------------------------------------------------------*/
 
-import client from './components/client';
-import provider from './components/provider';
-
-import dynamic from './routines/dynamic';
-import realtime from './routines/realtime';
-import tick from './routines/tick';
 import { EventEmitter } from 'events';
+import { Client } from './components/client';
+import { Provider } from './components/provider';
 
-import { ProviderConfig, Provider, ClientConfig, Client } from '../../../types';
+import { dynamic } from './routines/dynamic';
+import { realtime } from './routines/realtime';
+import { tick } from './routines/tick';
 
 /* Local variables -----------------------------------------------------------*/
 
@@ -22,24 +20,18 @@ const defaults = {
 
 /* Methods -------------------------------------------------------------------*/
 
-function listen(options: ProviderConfig): Provider {
-  options.label = options.label || Math.random().toString(36).substring(7);
-  return provider({ ...defaults, ...options }, new EventEmitter());
+const uniqueLabel = () => Math.random().toString(36).substring(7);
+
+export function listen(options: ProviderConfig): Provider {
+  return Provider({ label: uniqueLabel(), ...defaults, ...options }, new EventEmitter());
 }
 
-function connect(options: ClientConfig): Client {
-  options.label = options.label || Math.random().toString(36).substring(7);
-  return client({ ...defaults, ...options }, new EventEmitter());
+export function connect(options: ClientConfig): Client {
+  return Client({ label: uniqueLabel(), ...defaults, ...options }, new EventEmitter());
 }
 
-/* Exports -------------------------------------------------------------------*/
-
-export default {
-  connect,
-  listen,
-  routines: {
-    dynamic,
-    realtime,
-    tick,
-  },
+export const routines = {
+  dynamic,
+  realtime,
+  tick,
 };
