@@ -1,33 +1,16 @@
 /* Requires ------------------------------------------------------------------*/
 
 import dgram from 'dgram';
-import {
-  Socket,
-  Transport,
-  ClientConfig,
-  Remote,
-  SocketHandle,
-  UDPSocketHandle,
-  UDPConfig,
-  UDPClientList,
-  UDPClient,
-} from '../../../types';
-import { EventEmitter } from 'events';
 
 /* Methods -------------------------------------------------------------------*/
 
-function udp({
-  type = 'udp4',
-  localAddr = '0.0.0.0',
-  reuseAddr = true,
-  socketTimeout = 30000,
-  connectTimeout = 1000,
-}: UDPConfig = {}): Transport {
-  return function socket(params: ClientConfig, emitter: EventEmitter): Socket {
+function udp({ type = 'udp4', localAddr = '0.0.0.0', reuseAddr = true, socketTimeout = 30000, connectTimeout = 1000,
+}: UDPConfig = {}): KalmTransport {
+  return function socket(params: ClientConfig, emitter: NodeJS.EventEmitter): Socket {
     let listener: dgram.Socket;
     const clientCache: UDPClientList = {};
 
-    function addClient(client) {
+    function addClient(client: Client): void {
       const local: Remote = client.local();
       const key: string = `${local.host}.${local.port}`;
 
@@ -40,45 +23,6 @@ function udp({
         clientCache[key].client.emit('frame', clientCache[key].data[i]);
       }
       clientCache[key].data.length = 0;
-    }
-
-    function resolveClient(origin, data) {
-      const handle: UDPSocketHandle = {
-        host: origin.address,
-        port: origin.port,
-        socket: listener,
-      };
-
-      // Handle SYN
-      if (data[0] === 83 && data[1] === 89 && data[2] === 78) {
-        send(handle, Buffer.from('ACK'));
-        data = null;
-      }
-
-      const key: string = `${origin.address}.${origin.port}`;
-      clearTimeout(clientCache[key] && clientCache[key].timeout);
-
-      if (!clientCache[key]) {
-        clientCache[key] = {
-          client: null,
-          data: [],
-          timeout: null,
-        } as UDPClient;
-        emitter.emit('socket', handle);
-      }
-
-      if (data) {
-        if (clientCache[key].client) clientCache[key].client.emit('frame', data);
-        else clientCache[key].data.push(data);
-      }
-    }
-
-    function bind(): void {
-      listener = dgram.createSocket({ type: type as dgram.SocketType, reuseAddr });
-      listener.on('message', (data, origin) => resolveClient(origin, data));
-      listener.on('error', err => emitter.emit('error', err));
-      listener.bind(params.port, localAddr);
-      emitter.emit('ready');
     }
 
     function remote(handle: SocketHandle): Remote {
@@ -95,6 +39,11 @@ function udp({
       listener.close();
     }
 
+    function disconnect(handle?: UDPSocketHandle): void {
+      if (handle && handle.socket) handle.socket = null;
+      emitter.emit('disconnect');
+    }
+
     function connect(handle?: SocketHandle): SocketHandle {
       if (handle) return handle;
       const connection = dgram.createSocket(type as dgram.SocketType);
@@ -105,7 +54,7 @@ function udp({
         if (req[0] === 65 && req[1] === 67 && req[2] === 75) {
           clearTimeout(timeout);
           emitter.emit('connect', connection);
-        } else emitter.emit('frame', req);
+        } else emitter.emit('rawFrame', req);
       });
       connection.bind(null, localAddr);
 
@@ -122,9 +71,44 @@ function udp({
       return res;
     }
 
-    function disconnect(handle?: SocketHandle): void {
-      if (handle) handle = null;
-      emitter.emit('disconnect');
+    function resolveClient(origin, data) {
+      const handle: UDPSocketHandle = {
+        host: origin.address,
+        port: origin.port,
+        socket: listener,
+      };
+      let isSynPacket = false;
+
+      // Handle SYN
+      if (data[0] === 83 && data[1] === 89 && data[2] === 78) {
+        send(handle, Buffer.from('ACK'));
+        isSynPacket = true;
+      }
+
+      const key: string = `${origin.address}.${origin.port}`;
+      clearTimeout(clientCache[key] && clientCache[key].timeout);
+
+      if (!clientCache[key]) {
+        clientCache[key] = {
+          client: null,
+          data: [],
+          timeout: null,
+        } as UDPClient;
+        emitter.emit('socket', handle);
+      }
+
+      if (data && !isSynPacket) {
+        if (clientCache[key].client) clientCache[key].client.emit('rawFrame', data);
+        else clientCache[key].data.push(data);
+      }
+    }
+
+    function bind(): void {
+      listener = dgram.createSocket({ type: type as dgram.SocketType, reuseAddr });
+      listener.on('message', (data, origin) => resolveClient(origin, data));
+      listener.on('error', err => emitter.emit('error', err));
+      listener.bind(params.port, localAddr);
+      emitter.emit('ready');
     }
 
     emitter.on('connection', addClient);
@@ -142,4 +126,4 @@ function udp({
 
 /* Exports -------------------------------------------------------------------*/
 
-export default udp;
+module.exports = udp;
